feat: add API and binary
This commit adds the first version of the public API. It is subject to
change over time. The commit does also include a binary that can be used
to calculate the next semantic version of a repository.

The tool has to be run inside a git repository. If no specific revision
has been specified, the tool tries to find all git tags that match the
`vX.X.X` pattern. Those will then be used to determine the latest
semantic version of the repository. All commits from that tag on to the
current HEAD will be taken into account and used to calculate the next
version.

The tool does also offer a report functionality that prints out the
commits taken into account, their parsed results, as well as the range
the tool operated on. The reort can also be printed out to JSON to allow
for easier integration with scripts or other types of post-processing.